+++ source/main.ts
@@ -1,5 +1,5 @@
-export { Runner as CustomRunner } from "./engine/runners/runnerInterface.js";
+export { Runner as CustomRunner } from "./engine/task_runners/runnerInterface.js";
 export { TypeTasker } from "./engine/typeTasker.js";
-export { TypeTaskerCommand } from "./engine/runners/commandRunner.js";
-export { TypeTaskerEmpty } from "./engine/runners/emptyRunner.js";
-export { TypeTaskerCallback } from "./engine/runners/callbackRunner.js";
+export { TypeTaskerCommand } from "./engine/task_runners/commandRunner.js";
+export { TypeTaskerEmpty } from "./engine/task_runners/emptyRunner.js";
+export { TypeTaskerCallback } from "./engine/task_runners/callbackRunner.js";
